shinywaffle/requirements/requirements.txt
alabaster==0.7.9
appdirs==1.4.0
arrow==0.10.0
awesome-slugify==1.6.5
Babel==2.3.4
binaryornot==0.4.0
chardet==2.3.0
click==6.7
colorama==0.3.7
cookiecutter==1.5.1
coverage==4.3.4
decorator==4.0.11
defusedxml==0.5.0
Django==1.10.5
django-allauth==0.30.0
django-braces==1.11.0
django-coverage-plugin==1.4.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-environ==0.4.1
django-extensions==1.7.6
django-friendship==1.5.0
django-model-utils==2.6.1
django-redis==4.7.0
django-test-plus==1.0.17
docutils==0.13.1
factory-boy==2.8.1
Faker==0.7.7
future==0.16.0
imagesize==0.7.1
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
Jinja2==2.9.5
jinja2-time==0.2.0
MarkupSafe==0.23
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pickleshare==0.7.4
Pillow==4.0.0
poyo==0.4.0
prompt-toolkit==1.0.13
psycopg2==2.6.2
py==1.4.32
Pygments==2.2.0
pyparsing==2.1.10
pytest==3.0.6
pytest-django==3.1.2
pytest-sugar==0.8.0
python-dateutil==2.6.0
python-dotenv==0.6.2
python3-openid==3.0.10
pytz==2016.10
redis==2.10.5
regex==2017.2.8
requests==2.13.0
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.2
sqlparse==0.2.2
termcolor==1.1.0
traitlets==4.3.1
Unidecode==0.4.20
wcwidth==0.1.7
Werkzeug==0.11.15
whichcraft==0.4.0
